feat: scope GraphQL methods to admin
