+++ libs/admin/app/feature/src/index.ts
@@ -1,1 +1,2 @@
 export { AdminAppUserFeature as default } from './lib/admin-app-user-feature'
+export { AdminAppAdminFeatureList } from './lib/admin-app-admin-feature-list'
