+++ libs/admin/app/ui/src/lib/admin-app-ui-grid.tsx
@@ -1,14 +1,13 @@
-import { Box, Button, Code, Flex, SimpleGrid, Stack, Text } from '@chakra-ui/react'
+import { Flex, SimpleGrid, Stack, Text } from '@chakra-ui/react'
 import { App } from '@mogami/shared/util/admin-sdk'
 import React from 'react'
 import { Link } from 'react-router-dom'
 
 export interface AdminAppUiGridProps {
   apps: Partial<App>[] | null | undefined
-  deleteApp: (appId: string) => void
 }
 
-export function AdminAppUiGrid({ apps, deleteApp }: AdminAppUiGridProps) {
+export function AdminAppUiGrid({ apps }: AdminAppUiGridProps) {
   return (
     <SimpleGrid columns={2} gap={6}>
       {apps?.map((app) => (
@@ -21,13 +20,6 @@ export function AdminAppUiGrid({ apps, deleteApp }: AdminAppUiGridProps) {
             </Link>
             <Text fontSize="2xl">{app.index}</Text>
           </Flex>
-          <Box>
-            {app.id && (
-              <Button variant="ghost" color="red.500" size="xs" onClick={() => deleteApp(app.id!)}>
-                Delete
-              </Button>
-            )}
-          </Box>
         </Stack>
       ))}
     </SimpleGrid>
