+++ libs/admin/app/ui/src/lib/admin-app-ui-users.tsx
@@ -14,7 +14,7 @@ export function AdminAppUiUsers({ updateRole, users }: AdminAppUiUsersProps) {
     return <Alert>No users found.</Alert>
   }
   return (
-    <Box borderWidth="1px" borderRadius="lg" overflow="hidden" m="10px auto">
+    <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
       <TableContainer>
         <Table variant="simple" colorScheme="teal">
           <Thead>
