+++ libs/admin/ui/layout/src/lib/admin-ui-header-profile-menu.tsx
@@ -1,11 +1,12 @@
 import { Avatar, Button, Center, Menu, MenuButton, MenuDivider, MenuGroup, MenuItem, MenuList } from '@chakra-ui/react'
 import { AdminAuthLogoutFn } from '@mogami/admin/auth/data-access'
-import { User } from '@mogami/shared/util/admin-sdk'
+import { User, UserRole } from '@mogami/shared/util/admin-sdk'
 import React from 'react'
 import { Link } from 'react-router-dom'
 
 export function AdminUiHeaderProfileMenu({ logout, user }: { logout: AdminAuthLogoutFn; user: User }) {
   const defaultAvatar = 'https://avatars.githubusercontent.com/u/82999948?v=4'
+  const isAdmin = user.role === UserRole.Admin
   return (
     <Menu>
       <MenuButton as={Button} rounded={'full'} variant={'link'} cursor={'pointer'} minW={0}>
@@ -21,17 +22,22 @@ export function AdminUiHeaderProfileMenu({ logout, user }: { logout: AdminAuthLo
           <p>{user?.name}</p>
         </Center>
         <br />
-        <MenuGroup title="System Administration">
-          <MenuItem as={Link} to="/system/dashboard">
-            System Dashboard
-          </MenuItem>
-          <MenuItem as={Link} to="/system/clusters">
-            Manage Clusters
-          </MenuItem>
-          <MenuItem as={Link} to="/system/users">
-            Manage Users
-          </MenuItem>
-        </MenuGroup>
+        {isAdmin ? (
+          <MenuGroup title="System Administration">
+            <MenuItem as={Link} to="/system/dashboard">
+              Dashboard
+            </MenuItem>
+            <MenuItem as={Link} to="/system/apps">
+              Manage Apps
+            </MenuItem>
+            <MenuItem as={Link} to="/system/clusters">
+              Manage Clusters
+            </MenuItem>
+            <MenuItem as={Link} to="/system/users">
+              Manage Users
+            </MenuItem>
+          </MenuGroup>
+        ) : null}
         <MenuDivider />
         <MenuItem onClick={logout}>Logout</MenuItem>
       </MenuList>
