+++ libs/api/app/data-access/src/index.ts
@@ -1,3 +1,4 @@
+export * from './lib/api-app-admin-data-access.service'
 export * from './lib/api-app-data-access.module'
 export * from './lib/api-app-data-access.service'
 export * from './lib/api-app-user-data-access.service'
