+++ libs/api/app/data-access/src/lib/api-app-data-access.module.ts
@@ -2,13 +2,24 @@ import { ApiCoreDataAccessModule } from '@mogami/api/core/data-access'
 import { ApiWalletDataAccessModule } from '@mogami/api/wallet/data-access'
 import { HttpModule } from '@nestjs/axios'
 import { Module } from '@nestjs/common'
+import { ApiAppAdminDataAccessService } from './api-app-admin-data-access.service'
 import { ApiAppDataAccessService } from './api-app-data-access.service'
 import { ApiAppUserDataAccessService } from './api-app-user-data-access.service'
 import { ApiAppWebhookDataAccessService } from './api-app-webhook-data-access.service'
 
 @Module({
-  providers: [ApiAppDataAccessService, ApiAppWebhookDataAccessService, ApiAppUserDataAccessService],
-  exports: [ApiAppDataAccessService, ApiAppWebhookDataAccessService, ApiAppUserDataAccessService],
+  providers: [
+    ApiAppAdminDataAccessService,
+    ApiAppDataAccessService,
+    ApiAppWebhookDataAccessService,
+    ApiAppUserDataAccessService,
+  ],
+  exports: [
+    ApiAppAdminDataAccessService,
+    ApiAppDataAccessService,
+    ApiAppWebhookDataAccessService,
+    ApiAppUserDataAccessService,
+  ],
   imports: [ApiCoreDataAccessModule, ApiWalletDataAccessModule, HttpModule],
 })
 export class ApiAppDataAccessModule {}
