+++ libs/api/app/feature/src/lib/api-app-feature.controller.ts
@@ -1,4 +1,9 @@
-import { ApiAppDataAccessService, AppConfig, AppHealth } from '@mogami/api/app/data-access'
+import {
+  ApiAppDataAccessService,
+  ApiAppWebhookDataAccessService,
+  AppConfig,
+  AppHealth,
+} from '@mogami/api/app/data-access'
 import { Controller, Get, Param, ParseIntPipe, Post, Req, Res } from '@nestjs/common'
 import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger'
 import { Request, Response } from 'express'
@@ -6,20 +11,23 @@ import { Request, Response } from 'express'
 @ApiTags('app')
 @Controller('app')
 export class ApiAppFeatureController {
-  constructor(private readonly service: ApiAppDataAccessService) {}
+  constructor(
+    private readonly appService: ApiAppDataAccessService,
+    private readonly webhookService: ApiAppWebhookDataAccessService,
+  ) {}
 
   @Get(':environment/:index/config')
   @ApiOperation({ operationId: 'getAppConfig' })
   @ApiResponse({ type: AppConfig })
   app(@Param('environment') environment: string, @Param('index', ParseIntPipe) index: number) {
-    return this.service.getAppConfig(environment, index)
+    return this.appService.getAppConfig(environment, index)
   }
 
   @Get(':environment/:index/health')
   @ApiOperation({ operationId: 'getAppHealth' })
   @ApiResponse({ type: AppHealth })
   appHealth(@Param('environment') environment: string, @Param('index', ParseIntPipe) index: number) {
-    return this.service.getAppHealth(environment, index)
+    return this.appService.getAppHealth(environment, index)
   }
 
   @Post('/:environment/:index/webhook/:type')
@@ -30,6 +38,6 @@ export class ApiAppFeatureController {
     @Param('index', ParseIntPipe) index: number,
     @Param('type') type: string,
   ) {
-    return this.service.storeIncomingWebhook(environment, index, type, req.headers, req.body, res)
+    return this.webhookService.storeIncomingWebhook(environment, index, type, req.headers, req.body, res)
   }
 }
