+++ libs/api/app/feature/src/lib/api-app-feature.module.ts
@@ -1,5 +1,6 @@
 import { ApiAppDataAccessModule } from '@mogami/api/app/data-access'
 import { Module } from '@nestjs/common'
+import { ApiAppAdminFeatureResolver } from './api-app-admin-feature.resolver'
 import { ApiAppEnvFeatureResolver } from './api-app-env-feature.resolver'
 import { ApiAppFeatureController } from './api-app-feature.controller'
 import { ApiAppFeatureResolver } from './api-app-feature.resolver'
@@ -10,6 +11,7 @@ import { ApiAppUserFeatureResolver } from './api-app-user-feature.resolver'
   controllers: [ApiAppFeatureController],
   imports: [ApiAppDataAccessModule],
   providers: [
+    ApiAppAdminFeatureResolver,
     ApiAppEnvFeatureResolver,
     ApiAppFeatureResolver,
     ApiAppTransactionFeatureResolver,
